+++ src/matterbridge.ts
@@ -45,7 +45,7 @@ import { Frontend } from './frontend.js';
 
 // @matter
 import { DeviceTypeId, Endpoint as EndpointNode, Logger, LogLevel as MatterLogLevel, LogFormat as MatterLogFormat, VendorId, StorageContext, StorageManager, StorageService, Environment, ServerNode, FabricIndex, SessionsBehavior } from '@matter/main';
-import { DeviceCommissioner, ExposedFabricInformation, FabricAction, PaseClient } from '@matter/main/protocol';
+import { DeviceCommissioner, ExposedFabricInformation, FabricAction, MdnsService, PaseClient } from '@matter/main/protocol';
 import { AggregatorEndpoint } from '@matter/main/endpoints';
 
 // Default colors
@@ -222,9 +222,24 @@ export class Matterbridge extends EventEmitter {
    *
    */
   async destroyInstance() {
+    // Save server nodes to close
+    const servers: ServerNode<ServerNode.RootEndpoint>[] = [];
+    if (this.bridgeMode === 'bridge') {
+      if (this.serverNode) servers.push(this.serverNode);
+    }
+    if (this.bridgeMode === 'childbridge') {
+      for (const plugin of this.plugins.array()) {
+        if (plugin.serverNode) servers.push(plugin.serverNode);
+      }
+    }
+    // Cleanup
     await this.cleanup('destroying instance...', false);
-    // await matterServerNode.env.get(MdnsService)[Symbol.asyncDispose]();
-    // this.log.info(`Closed ${matterServerNode.id} MdnsService`);
+    // Close servers mdns service
+    for (const server of servers) {
+      await server.env.get(MdnsService)[Symbol.asyncDispose]();
+      this.log.info(`Closed ${server.id} MdnsService`);
+    }
+    // Wait for the cleanup to finish
     await new Promise((resolve) => {
       setTimeout(resolve, 1000);
     });
@@ -745,13 +760,13 @@ export class Matterbridge extends EventEmitter {
     process.removeAllListeners('unhandledRejection');
 
     this.exceptionHandler = async (error: Error) => {
-      this.log.fatal('Unhandled Exception detected at:', error.stack || error, rs);
+      this.log.error('Unhandled Exception detected at:', error.stack || error, rs);
       // await this.cleanup('Unhandled Exception detected, cleaning up...');
     };
     process.on('uncaughtException', this.exceptionHandler);
 
     this.rejectionHandler = async (reason, promise) => {
-      this.log.fatal('Unhandled Rejection detected at:', promise, 'reason:', reason instanceof Error ? reason.stack : reason, rs);
+      this.log.error('Unhandled Rejection detected at:', promise, 'reason:', reason instanceof Error ? reason.stack : reason, rs);
       // await this.cleanup('Unhandled Rejection detected, cleaning up...');
     };
     process.on('unhandledRejection', this.rejectionHandler);
@@ -1387,6 +1402,8 @@ export class Matterbridge extends EventEmitter {
       }
       this.hasCleanupStarted = false;
       this.initialized = false;
+    } else {
+      this.log.debug('Cleanup already started...');
     }
   }